refactor: error cause removed and favourite logic common out
